ui/requirements.txt
streamlit==1.38.0
requests==2.32.3
python-dotenv==1.0.1

# AWS SDK (UI only needs to read envs, no heavy boto3 usage, but keep it pinned)
boto3==1.34.162
botocore==1.34.162

# Pydantic is only needed if you parse configs using BaseModel
pydantic==2.8.2
pydantic-core==2.20.1
